pyproject.toml
[tool.poetry]
name = "podcast-storage"
version = "0.1.0"
description = ""
authors = ["williammadie <william.madie0@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.2"
pytest = "^8.3.4"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
